Searching overhaul.
I now trim the entries initially up to their day in order to reduce the amount of specific errors I get from classes with colons in their name or a letter sequence that matches the day. Although, yes, this does require more cycles to complete it solves a lot of problems in the future and makes the code more fault proof from classes that will be added in the future with colons. Just makes it so I don't have to comb through the registration looking for certain cases that break it.