Fixed time logic bug.
So I didn't actually fix the underlying problem but fixed it somewhere else, if that makes sense. The underlying bug is in the logic that shades in the entry. That is sorta wonky and I reused it for this so it made this weird. I do some other logic checks later on in the Classroom.java file that negate that bug and correct the display.